+++ app/controllers/items_controller.rb
@@ -12,12 +12,13 @@ class ItemsController < ApplicationController
   def create
     @item = Item.new(item_params)
     @item.user = current_user
+    puts item_params # 追加した行
     if @item.save
       redirect_to root_path, notice: "商品が出品されました。"
     else
-      render :new
+      render :new, status: :unprocessable_entity
     end
-  end  
+  end
 
   private
