+++ app/views/items/new.html.erb
@@ -1,5 +1,5 @@
 <head>
-<%= javascript_include_tag '/javascript/controllers/index.js' %>
+<script src = '/javascript/new.js'></script>
 </head>
 
 <div class="items-sell-contents">
@@ -9,11 +9,9 @@
   </header>
   <div class="items-sell-main">
     <h2 class="items-sell-title">商品の情報を入力</h2>
-    <%= form_with model: @item, local: true do |f| %>
 
-    <%# インスタンスを渡して、エラー発生時にメッセージが表示されるようにしましょう。%>
+    <%= form_with model: @item, local: true do |f| %>
     <%= render 'shared/error_messages', model: f.object %>
-    <%# //インスタンスを渡して、エラー発生時にメッセージが表示されるようにしましょう。%>
 
     <%# 商品画像 %>
     <div class="img-upload">
@@ -35,13 +33,13 @@
         商品名
         <span class="indispensable">必須</span>
       </div>
-     <%= f.text_field :name, class: "items-text", id: "item-name", placeholder: "商品名（必須 40文字まで)", maxlength: "40" %>
+     <%= f.text_field :name, class: "items-text", id: "item-name", placeholder: "商品名（必須 40文字まで)", maxlength: "40", value: @item.name %>
       <div class="items-explain">
         <div class="weight-bold-text">
           商品の説明
           <span class="indispensable">必須</span>
         </div>
-        <%= f.text_area :description, class: "items-text", id: "item-info", placeholder: "商品の説明（必須 1,000文字まで）（色、素材、重さ、定価、注意点など）例）2010年頃に1万円で購入したジャケットです。ライトグレーで傷はありません。あわせやすいのでおすすめです。" ,rows:"7" ,maxlength:"1000" %>
+        <%= f.text_area :description, class: "items-text", id: "item-info", placeholder: "商品の説明（必須 1,000文字まで）（色、素材、重さ、定価、注意点など）例）2010年頃に1万円で購入したジャケットです。ライトグレーで傷はありません。あわせやすいのでおすすめです。" ,rows:"7" ,maxlength:"1000",value: @item.description %>
       </div>
     </div>
     <%# /商品名と商品説明 %>
@@ -54,12 +52,13 @@
           カテゴリー
           <span class="indispensable">必須</span>
         </div>
-       <%= f.collection_select(:category_id, Category.all, :id, :name, {}, {class: "select-box", id: "item-category"}) %>
+      <%= f.collection_select(:category_id, Category.all, :id, :name, {}, { class: "select-box", id: "item-category" }) %>
+
         <div class="weight-bold-text">
           商品の状態
           <span class="indispensable">必須</span>
         </div>
-        <%= f.collection_select(:condition_id, Condition.all, :id, :name, {}, {class: "select-box", id: "item-sales-status"}) %>
+        <%= f.collection_select(:condition_id, Condition.all, :id, :name, {}, {class: "select-box", id: "item_condition"}) %>
       </div>
     </div>
     <%# /商品の詳細 %>
@@ -85,7 +84,7 @@
           発送までの日数
           <span class="indispensable">必須</span>
         </div>
-        <%= f.collection_select :shipping_day_id, ShippingDay.all, :id, :name, { prompt: "---" }, class: "select-box", id: "item-scheduled-delivery" %>
+        <%= f.collection_select(:shipping_day_id, ShippingDay.all, :id, :name, {}, {class: "select-box", id: "item-scheduled-delivery"}) %>
       </div>
     </div>
     <%# /配送について %>
